Add new configuration option: channel_scan_rounds=X
This configuration option determines the number of times horst scans the channel
spectrum when channel_scan=1. By default, channel_scan_rounds=-1 which means
that horst scans forever, so this commit preserves the old default behavior.

The purpose of this configuration option is to make it possible to run horst for
a predetermined period of time. If used in conjunction with channel=,
channel_upper= and channel_dwell=, horst can be configured to scan an arbitrary
range of channels for an arbitrary but predetermined period of time.

This allows horst to be used as a generic, well-controlled and easy-to-use
"backend" process providing scanning results for higher-level applications.