Do not call unsafe functions from signal handlers
POSIX defines a list of functions which can safely be interrupted by signals and
re-entered by signal handlers [1]. However, exit() is not included (at least
partially because exit() cannot guarantee functions registered with atexit() are
safe). Calling unsafe functions from signal handlers causes undefined behavior
(in practice deadlocks, heap corruptions etc.).

This commit makes sigint_handler() safe by making it set an atomic flag instead
of calling exit(). The flag is then tested in the mainloop and exit() is called
if the flag was set. Effectively, the burden of signal handling is moved from
the "interrupt context" to the "process context".

Furthermore, a standard pattern of signal masking + pselect() is applied to
avoid race conditions in the signal flag management.

[1]: http://man7.org/linux/man-pages/man7/signal.7.html