New func: ifctrl_iwadd_monitor(): add virtual 802.11 monitor interfaces
Function ifctrl_iwadd_monitor() is the first function in the new internal
interface control API declared in ifctrl.h. Currently, there's only a
Linux-specific implementation (using Netlink), ifctrl-nl80211.c, but the
ifctrl.h is designed to not expose any implementation details to make it easy to
switch the implementation on build time. So, if someone comes up with an
implementation for another platform, say OSX for example, she would need to just
write and compile OSX-specific module, and then link horst with it instead of
ifctrl-nl80211.o.