Exit with an error if quiet and initial channel selection fails
In UI mode, the user can just continue doing other things even if the
initial channel selection fails, but in quiet mode, failing to do as
told should be considered as a fatal error. There's no point in
continuing if horst fails to meet the configuration requirements
already in the initialization phase.