Standardize to bool return codes for success and failure
Standardize return codes for functions which return success or failure
to bool: success = true, failure = false.

I know that in C and glibc historically -1 and 0 are used for that usually
but to me that seems counter-intuitive and using bool as a return type makes it
more clear and easy to use. Most of the time the errors are handled or printed
inside the function and from the outside we are not interested in the sepecific
error code, rather we want to know "success or failure".

Most of the functions in horst were already that way, and for the ifctrl
functions the semantics changed.

Also made some fucntions return void.