More nl80211 fixes
As some commands return a reply message even though we just ignore it (e.g.:
add monitor interface), its better to always use the nl80211_send_recv() method
with an default empty message handler. Otherwise the reply message gets stuck
and then the next command gets confused. The effect of this was that the
frequency list was empty after we added a new monitor interface. We keep
nl80211_send() for convenience.

Also in order to get correct error messages for nl_perror we need to convert
the error code in the error callback, similar to how it's done in the default
verbose error handler of libnl.