Extend channel list structure to include band information
Add "band" (2.4GHz, 5GHz band) structure and define one struct to keep all
channels and bands which we can pass to ifctll to get the channels and band
definitions. I keep the two-dimensional array for the channels so I don't have
to change too much in the channel code. For now just keep the number of
channels in each band, but we will extend it later to keep more band
information.

Since the channel_list structure and num_channels is now "local" to channel.c
provide getter and setter functions.