Set maximum channel widths
Always try to set the maximum supported channel width on the monitor interface.
Otherwise, e.g. when the monitor interface is in HT20 mode (20MHz channel
width) we will not see HT40 (40MHz) data traffic.

For HT40 we'll have to visit each channel twice, once with HT40- and once with
HT40+.

TODO:
 - extend channel screen (display)
 - extend network protocol