Exit gracefully if conf.intf.ifname does not exist
horst will only work if there is a wlan interface. But if any user
tries horst on a system where there is no wifi hardware then it results
in a segfault.
As an added benifit this will also protect the user against any
misconfiguration of the wlan interface name.

The segfault has been reported in Ubuntu.

Signed-off-by: Sudip Mukherjee <sudipm.mukherjee@gmail.com>